backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
backports-datetime-fromisoformat==2.0.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.0.1
dj-email-url==1.0.6
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.1
django-cache-url==3.4.5
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.28.0
environs==14.3.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
marshmallow==4.0.1
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
python3-openid==3.2.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
whitenoise==6.11.0
